+++ src/ir/abstract.h
@@ -304,6 +304,33 @@ inline BinaryOp getBinary(Type type, Op op) {
   WASM_UNREACHABLE("invalid type");
 }
 
+inline Op negateRelational(Op op) {
+  switch (op) {
+    case Eq:
+      return Ne;
+    case Ne:
+      return Eq;
+    case LtS:
+      return GeS;
+    case LtU:
+      return GeU;
+    case LeS:
+      return GtS;
+    case LeU:
+      return GtU;
+    case GtS:
+      return LeS;
+    case GtU:
+      return LeU;
+    case GeS:
+      return LtS;
+    case GeU:
+      return LtU;
+    default:
+      WASM_UNREACHABLE("invalid relational");
+  }
+}
+
 } // namespace wasm::Abstract
 
 #endif // wasm_ir_abstract_h